Rename some headers
